requirements.txt
google-analytics-data>=0.18.2
pandas>=2.2.0
python-dotenv>=1.0.0
numpy>=2.0.0

